sirastudio_ai/requirements.txt
django>=6.0,<7.0
django-ninja>=1.6,<2.0
django-cors-headers>=4.0,<5.0
langchain>=1.0,<2.0
langgraph>=1.1,<2.0
langgraph-checkpoint-sqlite>=3.0,<4.0
langchain-openrouter>=0.2,<1.0
pydantic>=2.0,<3.0
python-dotenv>=1.0,<2.0
